Move property mapping generators to their own namespace.
